requirements.txt
# Core dependencies
streamlit>=1.12.0
python-dotenv>=1.0.0
altair>=4.2.0

# PDF processing
PyMuPDF>=1.21.1
PyPDF2>=3.0.0
python-docx>=0.8.11
docx2pdf>=0.1.8
streamlit_sortables>=0.3.1
pdf2docx>=0.5.8

# Firebase
firebase-admin>=6.0.0
pyrebase4>=4.5.0

# Image processing
Pillow>=9.0.0

# Country data
pycountry>=22.3.5

# Date/time
python-dateutil>=2.8.2
pytz>=2022.1

# Optional (if needed for your firebase_config)
google-auth>=2.0.0
google-cloud-firestore>=2.0.0

# Fix pkg_resources error
setuptools>=68.0.0
